tests/requirements.txt
asn1crypto==0.23.0
certifi==2017.7.27.1
cffi==1.11.1
chardet==3.0.4
cryptography==2.0.3
decorator==4.1.2
Django==1.11.5
django-rest-framework==0.1.0
djangorestframework==3.6.4
dnspython==1.15.0
enum34==1.1.6
freeipa==4.5.2
gssapi==1.2.0
httpie==0.9.9
idna==2.6
ipaclient==4.5.2
ipaddress==1.0.18
ipalib==4.5.2
ipapython==4.5.2
Jinja2==2.9.6
MarkupSafe==1.0
netaddr==0.7.19
netifaces==0.10.6
psycopg2==2.7.3.1
pyasn1==0.3.7
pyasn1-modules==0.1.4
pycparser==2.18
Pygments==2.2.0
python-ldap==2.4.44
python-nss==1.0.1
pytz==2017.2
PyYAML==3.12
pyzabbix==0.7.4
qrcode==5.3
requests==2.18.4
six==1.11.0
urllib3==1.22
xmltodict==0.11.0
